+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Revision history for electoral-sim
 
+* Creation of a random map of votes added 
+
 ## 0.1.0.0 -- 2020-06-21
 
 * First version. Released on an unsuspecting world.